Removed StaticEmailAlertSender, merged functionality into Formatted...
