+++ graylog2-server/src/main/java/org/graylog2/alerts/FormattedEmailAlertSender.java
@@ -17,24 +17,41 @@
 package org.graylog2.alerts;
 
 import com.floreysoft.jmte.Engine;
+import com.google.common.annotations.VisibleForTesting;
+import com.google.common.base.Strings;
+import org.apache.commons.mail.DefaultAuthenticator;
+import org.apache.commons.mail.Email;
+import org.apache.commons.mail.EmailConstants;
+import org.apache.commons.mail.EmailException;
+import org.apache.commons.mail.SimpleEmail;
 import org.graylog2.configuration.EmailConfiguration;
+import org.graylog2.notifications.Notification;
 import org.graylog2.notifications.NotificationService;
 import org.graylog2.plugin.Message;
+import org.graylog2.plugin.Tools;
 import org.graylog2.plugin.alarms.AlertCondition;
+import org.graylog2.plugin.alarms.transports.TransportConfigurationException;
 import org.graylog2.plugin.configuration.Configuration;
 import org.graylog2.plugin.streams.Stream;
 import org.graylog2.plugin.system.NodeId;
-import org.graylog2.streams.StreamRuleService;
+import org.joda.time.DateTime;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 
 import javax.inject.Inject;
+import java.net.URI;
 import java.util.Collections;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
+import java.util.Set;
 
 import static com.google.common.base.MoreObjects.firstNonNull;
+import static com.google.common.base.Strings.isNullOrEmpty;
 
-public class FormattedEmailAlertSender extends StaticEmailAlertSender implements AlertSender {
+public class FormattedEmailAlertSender implements AlertSender {
+    private static final Logger LOG = LoggerFactory.getLogger(FormattedEmailAlertSender.class);
+
     public static final String bodyTemplate = "##########\n" +
             "Alert Description: ${check_result.resultDescription}\n" +
             "Date: ${check_result.triggeredAt}\n" +
@@ -57,24 +74,25 @@ public class FormattedEmailAlertSender extends StaticEmailAlertSender implements
             "\n";
 
     private final Engine engine = new Engine();
+    private final NotificationService notificationService;
+    private final NodeId nodeId;
     private Configuration pluginConfig;
 
+    private final EmailConfiguration configuration;
+
     @Inject
-    public FormattedEmailAlertSender(EmailConfiguration configuration,
-                                     StreamRuleService streamRuleService,
-                                     NotificationService notificationService,
-                                     NodeId nodeId) {
-        super(configuration, streamRuleService, notificationService, nodeId);
+    public FormattedEmailAlertSender(EmailConfiguration configuration, NotificationService notificationService, NodeId nodeId) {
+        this.configuration = configuration;
+        this.notificationService = notificationService;
+        this.nodeId = nodeId;
     }
 
-    @Override
     public void initialize(Configuration configuration) {
         this.pluginConfig = configuration;
-        super.initialize(configuration);
     }
 
-    @Override
-    protected String buildSubject(Stream stream, AlertCondition.CheckResult checkResult, List<Message> backlog) {
+    @VisibleForTesting
+    String buildSubject(Stream stream, AlertCondition.CheckResult checkResult, List<Message> backlog) {
         final String template;
         if (pluginConfig == null || pluginConfig.getString("subject") == null) {
             template = "Graylog alert for stream: ${stream.title}: ${check_result.resultDescription}";
@@ -88,8 +106,8 @@ public class FormattedEmailAlertSender extends StaticEmailAlertSender implements
         return engine.transform(template, model);
     }
 
-    @Override
-    protected String buildBody(Stream stream, AlertCondition.CheckResult checkResult, List<Message> backlog) {
+    @VisibleForTesting
+    String buildBody(Stream stream, AlertCondition.CheckResult checkResult, List<Message> backlog) {
         final String template;
         if (pluginConfig == null || pluginConfig.getString("body") == null) {
             template = bodyTemplate;
@@ -114,4 +132,95 @@ public class FormattedEmailAlertSender extends StaticEmailAlertSender implements
 
         return model;
     }
+
+    private String buildStreamDetailsURL(URI baseUri, AlertCondition.CheckResult checkResult, Stream stream) {
+        // Return an informational message if the web interface URL hasn't been set
+        if (baseUri == null || isNullOrEmpty(baseUri.getHost())) {
+            return "Please configure 'transport_email_web_interface_url' in your Graylog configuration file.";
+        }
+
+        int time = 5;
+        if (checkResult.getTriggeredCondition().getParameters().get("time") != null) {
+            time = (int) checkResult.getTriggeredCondition().getParameters().get("time");
+        }
+
+        DateTime dateAlertEnd = checkResult.getTriggeredAt();
+        DateTime dateAlertStart = dateAlertEnd.minusMinutes(time);
+        String alertStart = Tools.getISO8601String(dateAlertStart);
+        String alertEnd = Tools.getISO8601String(dateAlertEnd);
+
+        return baseUri + "/streams/" + stream.getId() + "/messages?rangetype=absolute&from=" + alertStart + "&to=" + alertEnd + "&q=*";
+    }
+
+    public void sendEmails(Stream stream, EmailRecipients recipients, AlertCondition.CheckResult checkResult) throws TransportConfigurationException, EmailException {
+        sendEmails(stream, recipients, checkResult, null);
+    }
+
+    private void sendEmail(String emailAddress, Stream stream, AlertCondition.CheckResult checkResult, List<Message> backlog) throws TransportConfigurationException, EmailException {
+        LOG.debug("Sending mail to " + emailAddress);
+        if(!configuration.isEnabled()) {
+            throw new TransportConfigurationException("Email transport is not enabled in server configuration file!");
+        }
+
+        final Email email = new SimpleEmail();
+        email.setCharset(EmailConstants.UTF_8);
+
+        if (Strings.isNullOrEmpty(configuration.getHostname())) {
+            throw new TransportConfigurationException("No hostname configured for email transport while trying to send alert email!");
+        } else {
+            email.setHostName(configuration.getHostname());
+        }
+        email.setSmtpPort(configuration.getPort());
+        if (configuration.isUseSsl()) {
+            email.setSslSmtpPort(Integer.toString(configuration.getPort()));
+        }
+
+        if(configuration.isUseAuth()) {
+            email.setAuthenticator(new DefaultAuthenticator(
+                Strings.nullToEmpty(configuration.getUsername()),
+                Strings.nullToEmpty(configuration.getPassword())
+            ));
+        }
+
+        email.setSSLOnConnect(configuration.isUseSsl());
+        email.setStartTLSEnabled(configuration.isUseTls());
+        if (pluginConfig != null && !Strings.isNullOrEmpty(pluginConfig.getString("sender"))) {
+            email.setFrom(pluginConfig.getString("sender"));
+        } else {
+            email.setFrom(configuration.getFromEmail());
+        }
+        email.setSubject(buildSubject(stream, checkResult, backlog));
+        email.setMsg(buildBody(stream, checkResult, backlog));
+        email.addTo(emailAddress);
+
+        email.send();
+    }
+
+    @Override
+    public void sendEmails(Stream stream, EmailRecipients recipients, AlertCondition.CheckResult checkResult, List<Message> backlog) throws TransportConfigurationException, EmailException {
+        if(!configuration.isEnabled()) {
+            throw new TransportConfigurationException("Email transport is not enabled in server configuration file!");
+        }
+
+        if (recipients == null || recipients.isEmpty()) {
+            throw new RuntimeException("Cannot send emails: empty recipient list.");
+        }
+
+        final Set<String> recipientsSet = recipients.getEmailRecipients();
+        if (recipientsSet.size() == 0) {
+            final Notification notification = notificationService.buildNow()
+                .addNode(nodeId.toString())
+                .addType(Notification.Type.GENERIC)
+                .addSeverity(Notification.Severity.NORMAL)
+                .addDetail("title", "Stream \"" + stream.getTitle() + "\" is alerted, but no recipients have been defined!")
+                .addDetail("description", "To fix this, go to the alerting configuration of the stream and add at least one alert recipient.");
+            notificationService.publishIfFirst(notification);
+        }
+
+        for (String email : recipientsSet) {
+            sendEmail(email, stream, checkResult, backlog);
+        }
+    }
 }
+
+
